Alterando Dominio

+++ src/Opea.Application/Commands/DeleteClientCommandHandler.cs
@@ -1,6 +1,5 @@
 ﻿using MediatR;
 using Opea.Domain.AggregatesModel.ClientAggregate;
-using Opea.Domain.Events;
 
 namespace Opea.Application.Commands
 {
@@ -15,9 +14,13 @@ namespace Opea.Application.Commands
 
         public async Task<bool> Handle(DeleteClientCommand request, CancellationToken cancellationToken)
         {
-            var client = new Client(request.Id, request.CompanyName, request.CompanySizeId);
+            var client = await _clientRepository.GetByIdAsync(request.Id);
 
-            client.AddDomainEvent(new ClientRemovedEvent(client.Id, client.CompanyName, client.CompanySizeId));
+            if (client is null)
+                return false;
+
+            client.Delete(request.Id);
+
             _clientRepository.Delete(client);    
 
             return await _clientRepository.UnitOfWork.SaveEntitiesAsync(cancellationToken);
